starting to refactor code and preparing for automated test execution
